Struct wasmer_wasm_interface::interface_matcher::InterfaceMatcher

pub struct InterfaceMatcher {
    pub namespaces: HashSet<String>,
    pub namespace_imports: HashMap<String, HashSet<Import>>,
    pub exports: HashSet<Export>,
}

A struct containing data for more efficient matching.

An ideal use case for this is to parse [Interface]s at compile time, create InterfaceMatchers, and store them as bytes so that they can be efficiently loaded at runtime for matching.

Fields
namespaces: HashSet<String>namespace_imports: HashMap<String, HashSet<Import>>exports: HashSet<Export>