Module wasmer_wasm_interface::validate

Validate a wasm module given a interface.

This checks that all imports are specified in the interface and that their types are correct, as well as that all exports that the interface expects are exported by the module and that their types are correct.

Enums

WasmValidationError

Functions

validate_export_fns 🔒

Validates the export functions, checking the name and type against the given Interface

validate_export_globals 🔒

Validates the export globals, checking the name and type against the given Interface

validate_imports 🔒

Validates the import functions, checking the name and type against the given Interface

validate_wasm_and_report_errors

wasmparser_type_into_wasm_type 🔒

Converts Wasmparser’s type enum into wasm-interface’s type enum wasmparser’s enum contains things which are invalid in many situations